Create a single binary with server and cli commands
